% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse_codes.R
\name{tse_codes}
\alias{tse_codes}
\title{Return Numerical codes for Brazilian political parties from the
Tribunal Superior Eleitoral (TSE).}
\usage{
tse_codes()
}
\description{
Return Numerical codes for Brazilian political parties from the
Tribunal Superior Eleitoral (TSE).
}
